Address code review comments (create custom exception, move printing to the ParseState class, rethrow error).
