+++ LLama/Exceptions/GrammarFormatException.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LLama.Exceptions
+{
+    public class GrammarFormatException
+        : Exception
+    {
+        public GrammarFormatException()
+        {
+
+        }
+
+        public GrammarFormatException(string message)
+            : base(message)
+        {
+
+        }
+    }
+}
